# platforms/ycai/studio/docs/openapi-validated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "youchoose ai",
    "description": "youchoose api description",
    "version": "2.3.0"
  },
  "servers": [
    {
      "url": "{protocol}://{host}{port}{basePath}",
      "description": "Node Server",
      "variables": {
        "protocol": {
          "default": "https"
        },
        "host": {
          "default": "api.youchoose.ai"
        },
        "port": {
          "default": "443",
          "enum": ["443", "443"]
        },
        "basePath": {
          "default": ""
        }
      }
    }
  ],
  "security": [
    {
      "ACTToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ACTToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Authorization"
      }
    },
...
  },
  "paths": {
    "/v3/creator/ogp": {
      "post": {
        "summary": "CreateRecommendation",
        "description": "POST /v3/creator/ogp",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "CreateRecommendation",
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recommendation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Recommendation",
                  "properties": {
                    "urlId": {
                      "type": "string",
                      "description": "string"
                    },
                    "url": {
                      "type": "string",
                      "description": "string"
                    },
                    "title": {
                      "type": "string",
                      "description": "string"
                    },
                    "description": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "fromChannel": {
                      "type": "boolean",
                      "description": "A valid boolean type",
                      "required": false
                    },
                    "image": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/recommendations": {
      "get": {
        "summary": "CreatorRecommendations",
        "description": "GET /v3/creator/recommendations",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "RecommendationList",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "description": "RecommendationList[]",
                  "items": [
                    {
                      "type": "object",
                      "description": "Recommendation",
                      "properties": {
                        "urlId": {
                          "type": "string",
                          "description": "string"
                        },
                        "url": {
                          "type": "string",
                          "description": "string"
                        },
                        "title": {
                          "type": "string",
                          "description": "string"
                        },
                        "description": {
                          "type": "string",
                          "description": "string",
                          "required": false
                        },
                        "fromChannel": {
                          "type": "boolean",
                          "description": "A valid boolean type",
                          "required": false
                        },
                        "image": {
                          "type": "string",
                          "description": "string",
                          "required": false
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/{channelId}/related": {
      "get": {
        "summary": "CreatorRelatedChannels",
        "description": "GET /v3/creator/:channelId/related",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "integer32"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "integer32"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "GetRelatedChannelsOutput",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "GetRelatedChannelsOutput",
                  "properties": {
                    "content": {
                      "type": "array",
                      "description": "ChannelRelated[]",
                      "items": [
                        {
                          "type": "object",
                          "description": "ChannelRelated",
                          "properties": {
                            "recommendedSource": {
                              "type": "string",
                              "description": "string"
                            },
                            "recommendedChannelCount": {
                              "type": "number",
                              "format": "integer32"
                            },
                            "percentage": {
                              "type": "number",
                              "format": "integer32"
                            }
                          }
                        }
                      ]
                    },
                    "channelId": {
                      "type": "string",
                      "description": "string"
                    },
                    "authorName": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "score": {
                      "type": "number",
                      "format": "integer32"
                    },
                    "totalRecommendations": {
                      "type": "number",
                      "format": "integer32"
                    },
                    "totalContributions": {
                      "type": "number",
                      "format": "integer32"
                    },
                    "pagination": {
                      "type": "object",
                      "description": "{| amount: number |}",
                      "properties": {
                        "amount": {
                          "type": "number",
                          "format": "integer32"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/videos": {
      "get": {
        "summary": "CreatorVideos",
        "description": "GET /v3/creator/videos",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "ContentCreatorVideos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "description": "ContentCreatorVideos[]",
                  "items": [
                    {
                      "type": "object",
                      "description": "Video",
                      "properties": {
                        "description": {
                          "type": "string",
                          "description": "string",
                          "required": false
                        },
                        "recommendations": {
                          "type": "array",
                          "items": [
                            {
                              "type": "string"
                            }
                          ],
                          "required": false
                        },
                        "title": {
                          "type": "string",
                          "description": "string"
                        },
                        "videoId": {
                          "type": "string",
                          "description": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/recommendations/{urlId}": {
      "delete": {
        "summary": "DeleteRecommendation",
        "description": "DELETE /v3/creator/recommendations/:urlId",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          },
          {
            "name": "urlId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "boolean",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "description": "A `true | false` value"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "PatchRecommendation",
        "description": "PATCH /v3/creator/recommendations/:urlId",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          },
          {
            "name": "urlId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "PartialRecommendation",
                "properties": {
                  "urlId": {
                    "type": "string",
                    "description": "string"
                  },
                  "url": {
                    "type": "string",
                    "description": "string"
                  },
                  "title": {
                    "type": "string",
                    "description": "string"
                  },
                  "description": {
                    "type": "string",
                    "description": "string",
                    "required": false
                  },
                  "fromChannel": {
                    "type": "boolean",
                    "description": "A valid boolean type",
                    "required": false
                  },
                  "image": {
                    "type": "string",
                    "description": "string",
                    "required": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recommendation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Recommendation",
                  "properties": {
                    "urlId": {
                      "type": "string",
                      "description": "string"
                    },
                    "url": {
                      "type": "string",
                      "description": "string"
                    },
                    "title": {
                      "type": "string",
                      "description": "string"
                    },
                    "description": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "fromChannel": {
                      "type": "boolean",
                      "description": "A valid boolean type",
                      "required": false
                    },
                    "image": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/me": {
      "get": {
        "summary": "GetCreator",
        "description": "GET /v3/creator/me",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "ContentCreator",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "ContentCreator",
                  "properties": {
                    "channelId": {
                      "type": "string",
                      "description": "string"
                    },
                    "username": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "avatar": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "accessToken": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "url": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "registeredOn": {
                      "type": "object",
                      "description": "(undefined | DateFromISOString)",
                      "oneOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        }
                      ],
                      "required": false
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/{channelId}/stats": {
      "get": {
        "summary": "GetCreatorStats",
        "description": "GET /v3/creator/:channelId/stats",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "CreatorStats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "CreatorStats",
                  "properties": {
                    "authorName": {
                      "type": "string",
                      "description": "string"
                    },
                    "authorSource": {
                      "type": "string",
                      "description": "string"
                    },
                    "overflow": {
                      "type": "boolean",
                      "description": "A valid boolean type"
                    },
                    "total": {
                      "type": "number",
                      "format": "integer32"
                    },
                    "stripped": {
                      "type": "number",
                      "format": "integer32"
                    },
                    "content": {
                      "type": "array",
                      "description": "CreatorStatContent[]",
                      "items": [
                        {
                          "type": "object",
                          "description": "CreatorStatContent",
                          "properties": {
                            "id": {
                              "type": "string",
                              "description": "string"
                            },
                            "watchedTitle": {
                              "type": "string",
                              "description": "string"
                            },
                            "watchedVideoId": {
                              "type": "string",
                              "description": "string"
                            },
                            "savingTime": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "recommendedVideoId": {
                              "type": "string",
                              "description": "string"
                            },
                            "recommendedViews": {
                              "type": "number",
                              "format": "integer32"
                            },
                            "recommendedTitle": {
                              "type": "string",
                              "description": "string"
                            },
                            "recommendedChannel": {
                              "type": "string",
                              "description": "string"
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v3/creator/videos/{videoId}": {
      "get": {
        "summary": "OneCreatorVideo",
        "description": "GET /v3/creator/videos/:videoId",
        "tags": ["v3 - Creator"],
        "parameters": [
          {
            "name": "x-authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "description": "string"
            }
          },
          {
            "name": "videoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "string"
            }
          }
        ],
        "security": [
          {
            "ACTToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Video",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Video",
                  "properties": {
                    "description": {
                      "type": "string",
                      "description": "string",
                      "required": false
                    },
                    "recommendations": {
                      "type": "array",
                      "items": [
                        {
                          "type": "string"
                        }
                      ],
                      "required": false
                    },
                    "title": {
                      "type": "string",
                      "description": "string"
                    },
                    "videoId": {
                      "type": "string",
                      "description": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}